Refactor gui.py, add constants variable
