+++ gui/remove_window.py
@@ -1,0 +1,5 @@
+import tkinter as tk
+
+class remove_window(tk.Toplevel):
+    def __init__(self, mainWindow, root, obj):
+        print("will create the remove window")
