+++ gui/save_window.py
@@ -1,0 +1,5 @@
+import tkinter as tk
+
+class save_window(tk.Toplevel):
+    def __init__(self, mainWindow, root, obj):
+        print("will create the save window")
